move all queries to query folder
